Compiled project into one CS file, allowing compiling by bflat. Resulting binary appears to run on Xbox.
