--- KernelBase.cs
@@ -1,35 +1,0 @@
-﻿using Microsoft.Win32.SafeHandles;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace QuantumTunnel
-{
-    public static class KernelBase
-    {
-        [DllImport("kernelbase.dll", CharSet = CharSet.Auto, SetLastError = true)]
-        public static extern IntPtr CreateFile(
-        [MarshalAs(UnmanagedType.LPWStr)] string filename,
-        [MarshalAs(UnmanagedType.U4)] FileAccess access,
-        [MarshalAs(UnmanagedType.U4)] FileShare share,
-        IntPtr securityAttributes,
-        [MarshalAs(UnmanagedType.U4)] FileMode creationDisposition,
-        [MarshalAs(UnmanagedType.U4)] FileAttributes flagsAndAttributes,
-        IntPtr templateFile);
-
-        [DllImport("kernelbase.dll", SetLastError=true)]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool CloseHandle(IntPtr hObject);
-
-        [DllImport("kernelbase.dll", SetLastError = true)]
-        public static extern bool ReadFile(IntPtr hFile, [Out] byte[] lpBuffer,
-            uint nNumberOfBytesToRead, out uint lpNumberOfBytesRead, IntPtr lpOverlapped);
-
-        [DllImport("kernelbase.dll")]
-        public static extern uint GetLastError();
-    }
-}
